APIGateway/mobile-gateway/requirements.txt
Flask==2.2.0
requests==2.28.0
grpcio==1.51.0
grpcio-tools==1.51.0  